refactor: admin providers and portal providers + timeZone config
