+++ locales/en.json
@@ -1,5 +1,5 @@
 {
-  "search" : {
+  "search": {
     "search": "Search",
     "searchWelcome": "Welcome to your search page. This is where you can search for NFTs or minters.",
     "searchSentence": "Search an NFT or a minter.",
