+++ package.json
@@ -36,6 +36,7 @@
     "@radix-ui/react-navigation-menu": "^1.1.4",
     "@radix-ui/react-slot": "^1.0.2",
     "@tanstack/react-query": "^5.29.2",
+    "@types/intl": "^1.2.2",
     "@types/negotiator": "^0.6.3",
     "axios": "^1.6.8",
     "class-variance-authority": "^0.7.0",
@@ -53,6 +54,7 @@
     "sonner": "^1.4.41",
     "tailwind-merge": "^2.2.2",
     "tailwindcss-animate": "^1.0.7",
+    "use-intl": "^3.11.3",
     "vaul": "^0.9.0",
     "zod": "^3.22.4"
   },
